Update upper_layer3 and ns_convert for bgp-proc/as (#15)
* Update upper_layer3 and ns_convert for bgp-proc/as

* Refactoring: rename upper_layer3_filter

* Refactoring: convert tables

* Add module NamespaceConvertTable for convert table

---------
