+++ lib/api/topologies/network/snapshot/topology.rb
@@ -2,7 +2,7 @@
 
 require 'fileutils'
 require 'grape'
-require 'lib/convert_namespace/layer_filter'
+require 'lib/convert_namespace/upper_layer3_filter'
 require 'lib/topology_builder/topology_builder'
 require_relative 'topology/layer'
 require_relative 'topology/layer_type'
@@ -51,7 +51,7 @@ module NetomoxExp
         desc 'Get topology data (L3+ layers)'
         get 'upper_layer3' do
           network, snapshot = %i[network snapshot].map { |key| params[key] }
-          layer_filter = LayerFilter.new(read_topology_file(network, snapshot))
+          layer_filter = UpperLayer3Filter.new(read_topology_file(network, snapshot))
 
           # response
           layer_filter.filter
