+++ lib/convert_namespace/namespace_converter_base.rb
@@ -1,40 +1,78 @@
 # frozen_string_literal: true
 
-require 'json'
-
 module NetomoxExp
   # Base class of namespace converter
   class NamespaceConverterBase
-    # Target network (layer) names (regexp match)
-    TARGET_NW_REGEXP_LIST = [/ospf_area\d+/, /layer3/].freeze
+    # Target Network Types
+    UPPER_LAYER3_NWTYPE_LIST = [
+      Netomox::NWTYPE_MDDO_L3,
+      Netomox::NWTYPE_MDDO_OSPF_AREA,
+      Netomox::NWTYPE_MDDO_BGP_PROC,
+      Netomox::NWTYPE_MDDO_BGP_AS
+    ].freeze
 
     # Table of the keys which can not convert standard way (exceptional keys in L3/OSPF network)
     # Netomox::Topology attribute (object) -> Netomox::PseudoDSL attribute (Simple Hash)
     # NOTE: these keys are a list excepting `ip_addr`/`ip_address`
-    ATTR_KEY_TABLE = {
+    PLURAL_ATTR_KEY_TABLE = {
+      # plural + abbreviation key
+      ip_address: :ip_addrs,
+      # plural keys
       static_route: :static_routes,
       neighbor: :neighbors,
       prefix: :prefixes,
-      ip_address: :ip_addrs,
-      flag: :flags
+      flag: :flags,
+      confederation_member: :confederation_members,
+      peer_group: :peer_groups,
+      policy: :policies,
+      import_policy: :import_policies,
+      export_policy: :export_policies,
+      redistribute: :redistribute_list
     }.freeze
 
+    def initialize
+      # NOTE: initialized with #load_origin_topology
+      #   #convert_all_hash_keys and related methods are used in children: NamespaceConverter and UpperLayer3Filter.
+      #   Each class has different condition to initialize itself.
+      # @see NamespaceConvertTable#reload
+      #   It can make a instance of NamespaceConvertTable. It instance has two method to initialize:
+      #   1. Give it topology data (initialize from RFC8345 json)
+      #   2. Reload old convert table without topology data
+      @src_nws = nil
+    end
+
+    # @param [Hash] topology_data Topology data (RFC8345 Hash)
+    # @return [void]
+    def load_origin_topology(topology_data)
+      @src_nws = Netomox::Topology::Networks.new(topology_data)
+      @upper_l3_nw_names = upper_layer3_network_names
+    end
+
     protected
+
+    # @return [Array<String>] Network names (upper layer3)
+    def upper_layer3_network_names
+      nw_names = UPPER_LAYER3_NWTYPE_LIST.map do |network_type|
+        @src_nws&.find_all_networks_by_type(network_type)&.map(&:name)
+      end
+      nw_names.flatten.compact
+    end
 
     # @param [String] network_name Network (layer) name
     # @return [Boolean] True if the network_name matches one of TARGET_NW_REGEXP_LIST
     def target_network?(network_name)
-      TARGET_NW_REGEXP_LIST.any? { |nw_re| network_name =~ nw_re }
+      @upper_l3_nw_names.include?(network_name)
     end
 
     # @param [Symbol] key Key to convert
     # @param [Array, Object] value
     # @return [Symbol] Converted key
     def convert_hash_key(key, value)
-      # convert key symbol to snake_case
+      # convert key symbol (external key like 'os-type') to snake_case symbol (:os_type)
       converted_key = key.to_s.tr('-', '_').to_sym
-      return ATTR_KEY_TABLE[converted_key] if value.is_a?(Array) && ATTR_KEY_TABLE.key?(converted_key)
+      return PLURAL_ATTR_KEY_TABLE[converted_key] if value.is_a?(Array) && PLURAL_ATTR_KEY_TABLE.key?(converted_key)
 
+      # NOTE: irregular
       return :ip_addr if converted_key == :ip_address
 
       converted_key
